pyrig/pyproject.toml
[project]
name = "pyrig"
version = "0.1.0"
description = "Distributed device control for experimental rigs"
readme = "README.md"
license = { file = "../LICENSE" }
authors = [
    { name = "Walter Mwaniki", email = "walter.g.mwaniki@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.0",
    "pyzmq>=27.1.0",
    "ruyaml>=0.91.0",
    "vxlib",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
